# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Gestão de Carteira de Investimentos API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  ok: boolean;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface UserRegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface UserLoginInput {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
}

export type AssetCategory = typeof AssetCategory[keyof typeof AssetCategory];


export const AssetCategory = {
  acoes: 'acoes',
  fiis: 'fiis',
  etfs: 'etfs',
  bdrs: 'bdrs',
  fundos: 'fundos',
  renda_fixa: 'renda_fixa',
} as const;

export interface Asset {
  id: number;
  userId: number;
  ticker: string;
  quantity: number;
  averagePrice: number;
  /** @nullable */
  purchaseDate?: string | null;
  category: AssetCategory;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  currentPrice?: number | null;
  /** @nullable */
  totalValue?: number | null;
  /** @nullable */
  profitLoss?: number | null;
  /** @nullable */
  profitLossPercent?: number | null;
  createdAt: string;
}

export type AssetInputCategory = typeof AssetInputCategory[keyof typeof AssetInputCategory];


export const AssetInputCategory = {
  acoes: 'acoes',
  fiis: 'fiis',
  etfs: 'etfs',
  bdrs: 'bdrs',
  fundos: 'fundos',
  renda_fixa: 'renda_fixa',
} as const;

export interface AssetInput {
  /** @minLength 1 */
  ticker: string;
  quantity: number;
  averagePrice: number;
  /** @nullable */
  purchaseDate?: string | null;
  category: AssetInputCategory;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type AssetUpdateCategory = typeof AssetUpdateCategory[keyof typeof AssetUpdateCategory];


export const AssetUpdateCategory = {
  acoes: 'acoes',
  fiis: 'fiis',
  etfs: 'etfs',
  bdrs: 'bdrs',
  fundos: 'fundos',
  renda_fixa: 'renda_fixa',
} as const;

export interface AssetUpdate {
  ticker?: string;
  quantity?: number;
  averagePrice?: number;
  /** @nullable */
  purchaseDate?: string | null;
  category?: AssetUpdateCategory;
  /** @nullable */
  sector?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface PortfolioSummary {
  totalPatrimony: number;
  totalProfitLoss: number;
  totalProfitLossPercent: number;
  totalDividends: number;
  portfolioYield: number;
  assetCount: number;
}

export interface DistributionItem {
  label: string;
  value: number;
  percent: number;
}

export interface PortfolioDistribution {
  bySector: DistributionItem[];
  byCategory: DistributionItem[];
  byRisk: DistributionItem[];
}

export interface EvolutionPoint {
  date: string;
  value: number;
  /** @nullable */
  cdi?: number | null;
  /** @nullable */
  ibov?: number | null;
}

export type PortfolioHealthClassification = typeof PortfolioHealthClassification[keyof typeof PortfolioHealthClassification];


export const PortfolioHealthClassification = {
  Excelente: 'Excelente',
  Boa: 'Boa',
  Regular: 'Regular',
  Ruim: 'Ruim',
} as const;

export interface PortfolioHealth {
  score: number;
  classification: PortfolioHealthClassification;
  diversification: number;
  risk: number;
  dividends: number;
  growth: number;
  concentration: number;
}

export interface BenchmarkPoint {
  label: string;
  portfolio: number;
  cdi: number;
  ibov: number;
  ifix: number;
}

export interface BenchmarkComparison {
  points: BenchmarkPoint[];
  portfolioTotal: number;
  cdiTotal: number;
  ibovTotal: number;
  ifixTotal: number;
}

export type AlertType = typeof AlertType[keyof typeof AlertType];


export const AlertType = {
  preco: 'preco',
  fundamentos: 'fundamentos',
  dividendos: 'dividendos',
  noticias: 'noticias',
  macroeconomico: 'macroeconomico',
  concentracao: 'concentracao',
  substituicao: 'substituicao',
} as const;

export type AlertSeverity = typeof AlertSeverity[keyof typeof AlertSeverity];


export const AlertSeverity = {
  Critico: 'Critico',
  Alto: 'Alto',
  Medio: 'Medio',
  Baixo: 'Baixo',
} as const;

export interface Alert {
  id: number;
  userId: number;
  type: AlertType;
  severity: AlertSeverity;
  title: string;
  message: string;
  /** @nullable */
  ticker?: string | null;
  isRead: boolean;
  createdAt: string;
}

export type OpportunityRiskLevel = typeof OpportunityRiskLevel[keyof typeof OpportunityRiskLevel];


export const OpportunityRiskLevel = {
  Baixo: 'Baixo',
  Medio: 'Medio',
  Alto: 'Alto',
} as const;

export interface Opportunity {
  id: number;
  ticker: string;
  name: string;
  category: string;
  score: number;
  potentialReturn: number;
  dividendYield: number;
  riskLevel: OpportunityRiskLevel;
  reason: string;
  positives: string[];
  risks: string[];
  horizon: string;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  dividendo: 'dividendo',
  jcp: 'jcp',
  rendimento: 'rendimento',
  amortizacao: 'amortizacao',
} as const;

export interface Transaction {
  id: number;
  userId: number;
  ticker: string;
  amount: number;
  type: TransactionType;
  date: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  dividendo: 'dividendo',
  jcp: 'jcp',
  rendimento: 'rendimento',
  amortizacao: 'amortizacao',
} as const;

export interface TransactionInput {
  ticker: string;
  amount: number;
  type: TransactionInputType;
  date: string;
  /** @nullable */
  notes?: string | null;
}

export type AssetAnalysisStatus = typeof AssetAnalysisStatus[keyof typeof AssetAnalysisStatus];


export const AssetAnalysisStatus = {
  MANTER: 'MANTER',
  ATENCAO: 'ATENCAO',
  REAVALIAR: 'REAVALIAR',
  POSSIVEL_SAIDA: 'POSSIVEL_SAIDA',
} as const;

export type AssetAnalysisScoreClassification = typeof AssetAnalysisScoreClassification[keyof typeof AssetAnalysisScoreClassification];


export const AssetAnalysisScoreClassification = {
  Excelente: 'Excelente',
  Forte: 'Forte',
  Estavel: 'Estavel',
  Atencao: 'Atencao',
  Critico: 'Critico',
} as const;

export interface AssetAnalysis {
  ticker: string;
  status: AssetAnalysisStatus;
  score: number;
  scoreClassification: AssetAnalysisScoreClassification;
  positives: string[];
  risks: string[];
  newsItems: string[];
  alerts: string[];
  monitoringRecommendation: string;
  updatedAt: string;
}

export interface PortfolioReport {
  generatedAt: string;
  summary: PortfolioSummary;
  analyses: AssetAnalysis[];
  topAlerts: Alert[];
  opportunities: Opportunity[];
}