# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gestão de Carteira de Investimentos API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List portfolio assets
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Add an asset to portfolio
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get single asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetUpdate"
      responses:
        "200":
          description: Updated asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Remove an asset from portfolio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── PORTFOLIO SUMMARIES ─────────────────────────────────────────────────────
...
  /alerts/{id}/read:
    patch:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark an alert as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /alerts/{id}:
    delete:
      operationId: deleteAlert
      tags: [alerts]
      summary: Dismiss/delete an alert
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Alert deleted

  # ─── OPPORTUNITIES ───────────────────────────────────────────────────────────
...
  /analysis/assets/{ticker}:
    get:
      operationId: getAssetAnalysis
      tags: [analysis]
      summary: Get detailed analysis for a specific asset ticker
      parameters:
        - name: ticker
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Asset analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetAnalysis"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string
          format: date-time
    UserRegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
        password:
          type: string
          minLength: 6
    UserLoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # ─── ASSET SCHEMAS ────────────────────────────────────────────────────────
    Asset:
      type: object
      required: [id, userId, ticker, quantity, averagePrice, category, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        ticker:
          type: string
        quantity:
          type: number
        averagePrice:
          type: number
        purchaseDate:
          type: ["string", "null"]
          format: date
        category:
          type: string
          enum: [acoes, fiis, etfs, bdrs, fundos, renda_fixa]
        sector:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        currentPrice:
          type: ["number", "null"]
        totalValue:
          type: ["number", "null"]
        profitLoss:
          type: ["number", "null"]
        profitLossPercent:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    AssetInput:
      type: object
      required: [ticker, quantity, averagePrice, category]
      properties:
        ticker:
          type: string
          minLength: 1
        quantity:
          type: number
        averagePrice:
          type: number
        purchaseDate:
          type: ["string", "null"]
          format: date
        category:
          type: string
          enum: [acoes, fiis, etfs, bdrs, fundos, renda_fixa]
        sector:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    AssetUpdate:
      type: object
      properties:
        ticker:
          type: string
        quantity:
          type: number
        averagePrice:
          type: number
        purchaseDate:
          type: ["string", "null"]
          format: date
        category:
          type: string
          enum: [acoes, fiis, etfs, bdrs, fundos, renda_fixa]
        sector:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ─── PORTFOLIO SCHEMAS ────────────────────────────────────────────────────
...
    Alert:
      type: object
      required: [id, userId, type, severity, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [preco, fundamentos, dividendos, noticias, macroeconomico, concentracao, substituicao]
        severity:
          type: string
          enum: [Critico, Alto, Medio, Baixo]
        title:
          type: string
        message:
          type: string
        ticker:
          type: ["string", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # ─── OPPORTUNITY SCHEMAS ──────────────────────────────────────────────────
...
    AssetAnalysis:
      type: object
      required: [ticker, status, score, scoreClassification, positives, risks, newsItems, alerts, monitoringRecommendation, updatedAt]
      properties:
        ticker:
          type: string
        status:
          type: string
          enum: [MANTER, ATENCAO, REAVALIAR, POSSIVEL_SAIDA]
        score:
          type: number
        scoreClassification:
          type: string
          enum: [Excelente, Forte, Estavel, Atencao, Critico]
        positives:
          type: array
          items:
            type: string
        risks:
          type: array
          items:
            type: string
        newsItems:
          type: array
          items:
            type: string
        alerts:
          type: array
          items:
            type: string
        monitoringRecommendation:
          type: string
        updatedAt:
          type: string
          format: date-time
...