# lib/api-zod/src/generated/types/userRegisterInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Gestão de Carteira de Investimentos API
 * OpenAPI spec version: 0.1.0
 */

export interface UserRegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}